+++ HdbscanSharp/Runner/HdbscanResult.cs
@@ -8,6 +8,8 @@ namespace HdbscanSharp.Runner
 		public int[] Labels { get; set; }
 		public List<OutlierScore> OutliersScore { get; set; }
 		public bool HasInfiniteStability { get; set; }
+		public double RelativeValidity { get; set; }
+		public Dictionary<int, double> ClusterPersistence { get; set; }
 	}
 
 	public class HdbscanResult<T>
@@ -15,5 +17,7 @@ namespace HdbscanSharp.Runner
 		public Dictionary<int, List<T>> Groups { get; set; }
 		public List<OutlierScore<T>> OutliersScore { get; set; }
 		public bool HasInfiniteStability { get; set; }
+		public double RelativeValidity { get; set; }
+		public Dictionary<int, double> ClusterPersistence { get; set; }
 	}
 }